full release
for some WEIRD FUCKING REASON, 30% of kajigs weren't on the names list, but were very much present in the actual kajig info archives. meaning I did something wrong. anyways, fixed. also a little bit of a UX change for the viewer and now that every kajig is theoretically on this, kwi is full release